Modified plan for year 2.
I have been thinking about this (now is year 2, between D4 and D5 FT,
D2-D3 PT) and I think the solution to the connundrum

  - Trevor constraint, plus
  - Recursion contraint
  - Same content for both groups

is to break up the recursion days and have them apart instead of all
together. A bit of recursion at the beginning and then more recursion
later with the memoization and the divide-and-conquer; kind of like
"basic recursion" and "advanced recursion".