Committed the new file that contains only finals and javadoc.
I created it at home but forgot to add it, so it was not committed (or
pushed to pendrive). Now I have created it again and committed (and
will push to pendrive; I wonder what will happen when I pull at
home... will I get a 'are you sure to want to overwrite this local
file with the same name as the remote file we are pulling?' message?).