Updated ContactManager to make it clear that students must use TDD.
It used to be left to them, but experience shows that TDD is really
counterintuitive and they just prefer not to do it because they do not
understand it. This is wrong, and a disservice the faculty is doing to
the students. Given that we can analyse their version control history,
there is no excuse for not making TDD compulsory so they practice it
and learn it.